+++ CHANGELOG.md
@@ -9,6 +9,10 @@
   `javax.annotation.Generated` be present on the class path. The classes will only be annotated with
   this annotation if it is present on the classpath at the time of generation and the source version
   is `8`.
+* **\[processor\]** Enhance the annotation processor so that it will not attempt to process classes
+  until they are completely resolved. If an React4j component contains a dependency on code with compilation
+  errors or has a circular dependency with generated code then it will not able to be processed
+  by the new annotation processor.
 
 ### [v0.72](https://github.com/react4j/react4j/tree/v0.72) (2018-06-04)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.71...v0.72)